requirements.txt
openvino~=2023.0.1
numpy~=1.23.5
opencv-python~=4.6.0.66
gradio~=3.22.1
pillow~=9.4.0
ultralytics~=8.0.199
albumentations~=1.3.0
tqdm~=4.64.1
torch~=2.0.1+cu118
torchvision~=0.15.2+cu118